Fix `DEBUG=1` mode
 #### Summary
On my ubuntu 18.04 image, enabling the `DEBUG=1` flag to debug the
library fails to build with this error message:

```plaintext
meson.build:1:0: ERROR: Value true -Doptimization=g is not boolean (true or false).
```

Removing the single quotes enables the build to proceed!

 #### Testing
This patch doesn't add a build step to verify the `DEBUG=1` mode is
still working, so just tested locally!