Restore CRT symbol definitions.
Using a ** works on macOS when testing, but does not work on device - we need to adjust our logic in that case to take the address of the symbols (e.g., &__preinit_array_end). However, taking the address of these symbols makes the test fail again.

Not knowing whether this is a Clang or OS X specific behavior, we'll just use an #ifdef to select the proper form when testing.