Fix endianness declarations for ARM platform
The ARM platform previously referenced _LITTLE_ENDIAN and _BIG_ENDIAN
instead of __LITTLE_ENDIAN and __BIG_ENDIAN (which are what is defined
in endian.h and used for other platforms). This meant that ARM
platforms were treated as big-endian regardless of actual endianness.
By correcting the declarations to use the correct number of
underscores, ARM platforms can now be correctly recognized as
little-endian when __ARMEB__ is not defined.