Eliminate dangerous weak accesses to __assert_fail
Weak symbols have two functions: to permit overriding built-in
definitions with custom definitions, and to permit calling a function
even if it is not defined.

The former usage is useful for __assert_fail; it allows users to
safely override the definition. Unfortunately, the latter usage is
actively dangerous: when calling a function, a weak reference is
emitted, and if the particular linking order comes out differently than
expected, the weakly-referenced symbol may be omitted from the link
entirely. In this case, a NOP is inserted instead of a function call.

The dangerous impact of inserting a NOP in the place of __assert_fail
is that it is also marked as noreturn. This means that the compiler
will have generated code assuming that it would never return, and
placed arbitrary other code after this point in the binary output. This
silently prevents assertions from triggering (which may not be
immediately noticeable), and (even worse) causes the code that fails
those asserts to jump into random (and possibly unrelated) other pieces
of code.

In order to solve this, we simply need to remove the weak attribute
from the declaration of __assert_fail in the header file. The source
file can safely continue to carry this attribute, because that will
only affect the definition (allowing it to be overridden) without also
permitting it to go entirely undefined.