Convert from mpaland/printf to eyalroz/printf for the underlying embeddedartistry/printf implementation.
mpaland/printf has been sleepy, and eyalroz has been improving the printf implementation. This is the initial commit to switch the library implementation over.

Note that _putchar() is now expected to be putchar_() by the library, but we have created a _putchar() shim to maintain support for existing applications.

Fixes #159